TSP distance type changed from double to integer according to TSPLIB convention used in most benchmarks. TSPFitness changed to negative tour length for easy comparison of results with other benchmarks.
git-svn-id: svn://svn.code.sf.net/p/jcop/code/trunk@39 ca20af45-4a20-473b-94bf-0ef87cbc3059

+++ jcop/tests/cz/cvut/felk/cig/jcop/algorithm/geneticalgorithm/GeneticAlgorithmTest.java
@@ -40,7 +40,7 @@ public class GeneticAlgorithmTest {
         }
         try {
             geneticAlgorithm = new GeneticAlgorithm(10, 0.1);
-            geneticAlgorithm.init(new BaseObjectiveProblem(new TSP(new Double[][]{{1.0, 2.0}, {3.0, 4.0}})));
+            geneticAlgorithm.init(new BaseObjectiveProblem(new TSP(new Integer[][]{{1, 2}, {3, 4}})));
             assert false : "Expected exception, but none was risen.";
         } catch (InvalidProblemException ignore) {
